init lib build
