+++ packages/oas-swagger-ui/src/App.tsx
@@ -1,29 +1,13 @@
-// @ts-ignore
-import SwaggerUI from 'swagger-ui-react';
-import 'swagger-ui-react/swagger-ui.css';
-import { Operation } from './plugin-overrides/Operation.tsx';
-import { OasSwaggerUiConfig } from './plugin-overrides/types.ts';
+import { ComponentProps } from 'react';
+import { OasSwaggerUi } from './main-lib';
 
 function App({
   swaggerConfig,
   oasSwaggerUiConfig
-}: {
-  swaggerConfig: any;
-  oasSwaggerUiConfig?: OasSwaggerUiConfig;
-}) {
-  const plugins = {
-    wrapComponents: {
-      operation: () => (props: any) =>
-        (
-          <Operation
-            {...props}
-            oasSwaggerUiSecurityConfig={oasSwaggerUiConfig?.security}
-          />
-        )
-    }
-  };
-
-  return <SwaggerUI {...swaggerConfig} plugins={plugins} />;
+}: ComponentProps<typeof OasSwaggerUi>) {
+  return (
+    <OasSwaggerUi {...swaggerConfig} oasSwaggerUiConfig={oasSwaggerUiConfig} />
+  );
 }
 
 export default App;
